Release/s101

+++ tests/tsim/src/simParse.c
@@ -840,14 +840,14 @@ void simInitsimCmdList() {
   cmdno = SIM_CMD_EXP;
   simCmdList[cmdno].cmdno = cmdno;
   strcpy(simCmdList[cmdno].name, "exp");
-  simCmdList[cmdno].nlen = (int)strlen(simCmdList[cmdno].name);
+  simCmdList[cmdno].nlen = (int16_t)strlen(simCmdList[cmdno].name);
   simCmdList[cmdno].parseCmd = NULL;
   simCmdList[cmdno].executeCmd = simExecuteExpCmd;
 
   cmdno = SIM_CMD_IF;
   simCmdList[cmdno].cmdno = cmdno;
   strcpy(simCmdList[cmdno].name, "if");
-  simCmdList[cmdno].nlen = (int)strlen(simCmdList[cmdno].name);
+  simCmdList[cmdno].nlen = (int16_t)strlen(simCmdList[cmdno].name);
   simCmdList[cmdno].parseCmd = simParseIfCmd;
   simCmdList[cmdno].executeCmd = NULL;
   simAddCmdIntoHash(&(simCmdList[cmdno]));
@@ -855,7 +855,7 @@ void simInitsimCmdList() {
   cmdno = SIM_CMD_ELIF;
   simCmdList[cmdno].cmdno = cmdno;
   strcpy(simCmdList[cmdno].name, "elif");
-  simCmdList[cmdno].nlen = (int)strlen(simCmdList[cmdno].name);
+  simCmdList[cmdno].nlen = (int16_t)strlen(simCmdList[cmdno].name);
   simCmdList[cmdno].parseCmd = simParseElifCmd;
   simCmdList[cmdno].executeCmd = NULL;
   simAddCmdIntoHash(&(simCmdList[cmdno]));
@@ -863,7 +863,7 @@ void simInitsimCmdList() {
   cmdno = SIM_CMD_ELSE;
   simCmdList[cmdno].cmdno = cmdno;
   strcpy(simCmdList[cmdno].name, "else");
-  simCmdList[cmdno].nlen = (int)strlen(simCmdList[cmdno].name);
+  simCmdList[cmdno].nlen = (int16_t)strlen(simCmdList[cmdno].name);
   simCmdList[cmdno].parseCmd = simParseElseCmd;
   simCmdList[cmdno].executeCmd = NULL;
   simAddCmdIntoHash(&(simCmdList[cmdno]));
@@ -871,7 +871,7 @@ void simInitsimCmdList() {
   cmdno = SIM_CMD_ENDI;
   simCmdList[cmdno].cmdno = cmdno;
   strcpy(simCmdList[cmdno].name, "endi");
-  simCmdList[cmdno].nlen = (int)strlen(simCmdList[cmdno].name);
+  simCmdList[cmdno].nlen = (int16_t)strlen(simCmdList[cmdno].name);
   simCmdList[cmdno].parseCmd = simParseEndiCmd;
   simCmdList[cmdno].executeCmd = NULL;
   simAddCmdIntoHash(&(simCmdList[cmdno]));
@@ -879,7 +879,7 @@ void simInitsimCmdList() {
   cmdno = SIM_CMD_WHILE;
   simCmdList[cmdno].cmdno = cmdno;
   strcpy(simCmdList[cmdno].name, "while");
-  simCmdList[cmdno].nlen = (int)strlen(simCmdList[cmdno].name);
+  simCmdList[cmdno].nlen = (int16_t)strlen(simCmdList[cmdno].name);
   simCmdList[cmdno].parseCmd = simParseWhileCmd;
   simCmdList[cmdno].executeCmd = NULL;
   simAddCmdIntoHash(&(simCmdList[cmdno]));
@@ -887,7 +887,7 @@ void simInitsimCmdList() {
   cmdno = SIM_CMD_ENDW;
   simCmdList[cmdno].cmdno = cmdno;
   strcpy(simCmdList[cmdno].name, "endw");
-  simCmdList[cmdno].nlen = (int)strlen(simCmdList[cmdno].name);
+  simCmdList[cmdno].nlen = (int16_t)strlen(simCmdList[cmdno].name);
   simCmdList[cmdno].parseCmd = simParseEndwCmd;
   simCmdList[cmdno].executeCmd = NULL;
   simAddCmdIntoHash(&(simCmdList[cmdno]));
@@ -895,7 +895,7 @@ void simInitsimCmdList() {
   cmdno = SIM_CMD_SWITCH;
   simCmdList[cmdno].cmdno = cmdno;
   strcpy(simCmdList[cmdno].name, "switch");
-  simCmdList[cmdno].nlen = (int)strlen(simCmdList[cmdno].name);
+  simCmdList[cmdno].nlen = (int16_t)strlen(simCmdList[cmdno].name);
   simCmdList[cmdno].parseCmd = simParseSwitchCmd;
   simCmdList[cmdno].executeCmd = NULL;
   simAddCmdIntoHash(&(simCmdList[cmdno]));
@@ -903,7 +903,7 @@ void simInitsimCmdList() {
   cmdno = SIM_CMD_CASE;
   simCmdList[cmdno].cmdno = cmdno;
   strcpy(simCmdList[cmdno].name, "case");
-  simCmdList[cmdno].nlen = (int)strlen(simCmdList[cmdno].name);
+  simCmdList[cmdno].nlen = (int16_t)strlen(simCmdList[cmdno].name);
   simCmdList[cmdno].parseCmd = simParseCaseCmd;
   simCmdList[cmdno].executeCmd = NULL;
   simAddCmdIntoHash(&(simCmdList[cmdno]));
@@ -911,7 +911,7 @@ void simInitsimCmdList() {
   cmdno = SIM_CMD_DEFAULT;
   simCmdList[cmdno].cmdno = cmdno;
   strcpy(simCmdList[cmdno].name, "default");
-  simCmdList[cmdno].nlen = (int)strlen(simCmdList[cmdno].name);
+  simCmdList[cmdno].nlen = (int16_t)strlen(simCmdList[cmdno].name);
   simCmdList[cmdno].parseCmd = simParseDefaultCmd;
   simCmdList[cmdno].executeCmd = NULL;
   simAddCmdIntoHash(&(simCmdList[cmdno]));
@@ -919,7 +919,7 @@ void simInitsimCmdList() {
   cmdno = SIM_CMD_BREAK;
   simCmdList[cmdno].cmdno = cmdno;
   strcpy(simCmdList[cmdno].name, "break");
-  simCmdList[cmdno].nlen = (int)strlen(simCmdList[cmdno].name);
+  simCmdList[cmdno].nlen = (int16_t)strlen(simCmdList[cmdno].name);
   simCmdList[cmdno].parseCmd = simParseBreakCmd;
   simCmdList[cmdno].executeCmd = NULL;
   simAddCmdIntoHash(&(simCmdList[cmdno]));
@@ -927,7 +927,7 @@ void simInitsimCmdList() {
   cmdno = SIM_CMD_CONTINUE;
   simCmdList[cmdno].cmdno = cmdno;
   strcpy(simCmdList[cmdno].name, "continue");
-  simCmdList[cmdno].nlen = (int)strlen(simCmdList[cmdno].name);
+  simCmdList[cmdno].nlen = (int16_t)strlen(simCmdList[cmdno].name);
   simCmdList[cmdno].parseCmd = simParseContinueCmd;
   simCmdList[cmdno].executeCmd = NULL;
   simAddCmdIntoHash(&(simCmdList[cmdno]));
@@ -935,7 +935,7 @@ void simInitsimCmdList() {
   cmdno = SIM_CMD_ENDS;
   simCmdList[cmdno].cmdno = cmdno;
   strcpy(simCmdList[cmdno].name, "ends");
-  simCmdList[cmdno].nlen = (int)strlen(simCmdList[cmdno].name);
+  simCmdList[cmdno].nlen = (int16_t)strlen(simCmdList[cmdno].name);
   simCmdList[cmdno].parseCmd = simParseEndsCmd;
   simCmdList[cmdno].executeCmd = NULL;
   simAddCmdIntoHash(&(simCmdList[cmdno]));
@@ -943,7 +943,7 @@ void simInitsimCmdList() {
   cmdno = SIM_CMD_SLEEP;
   simCmdList[cmdno].cmdno = cmdno;
   strcpy(simCmdList[cmdno].name, "sleep");
-  simCmdList[cmdno].nlen = (int)strlen(simCmdList[cmdno].name);
+  simCmdList[cmdno].nlen = (int16_t)strlen(simCmdList[cmdno].name);
   simCmdList[cmdno].parseCmd = simParseSleepCmd;
   simCmdList[cmdno].executeCmd = simExecuteSleepCmd;
   simAddCmdIntoHash(&(simCmdList[cmdno]));
@@ -951,7 +951,7 @@ void simInitsimCmdList() {
   cmdno = SIM_CMD_GOTO;
   simCmdList[cmdno].cmdno = cmdno;
   strcpy(simCmdList[cmdno].name, "goto");
-  simCmdList[cmdno].nlen = (int)strlen(simCmdList[cmdno].name);
+  simCmdList[cmdno].nlen = (int16_t)strlen(simCmdList[cmdno].name);
   simCmdList[cmdno].parseCmd = simParseGotoCmd;
   simCmdList[cmdno].executeCmd = simExecuteGotoCmd;
   simAddCmdIntoHash(&(simCmdList[cmdno]));
@@ -959,7 +959,7 @@ void simInitsimCmdList() {
   cmdno = SIM_CMD_RUN;
   simCmdList[cmdno].cmdno = cmdno;
   strcpy(simCmdList[cmdno].name, "run");
-  simCmdList[cmdno].nlen = (int)strlen(simCmdList[cmdno].name);
+  simCmdList[cmdno].nlen = (int16_t)strlen(simCmdList[cmdno].name);
   simCmdList[cmdno].parseCmd = simParseRunCmd;
   simCmdList[cmdno].executeCmd = simExecuteRunCmd;
   simAddCmdIntoHash(&(simCmdList[cmdno]));
@@ -967,7 +967,7 @@ void simInitsimCmdList() {
   cmdno = SIM_CMD_RUN_BACK;
   simCmdList[cmdno].cmdno = cmdno;
   strcpy(simCmdList[cmdno].name, "run_back");
-  simCmdList[cmdno].nlen = (int)strlen(simCmdList[cmdno].name);
+  simCmdList[cmdno].nlen = (int16_t)strlen(simCmdList[cmdno].name);
   simCmdList[cmdno].parseCmd = simParseRunBackCmd;
   simCmdList[cmdno].executeCmd = simExecuteRunBackCmd;
   simAddCmdIntoHash(&(simCmdList[cmdno]));
@@ -975,7 +975,7 @@ void simInitsimCmdList() {
   cmdno = SIM_CMD_SYSTEM;
   simCmdList[cmdno].cmdno = cmdno;
   strcpy(simCmdList[cmdno].name, "system");
-  simCmdList[cmdno].nlen = (int)strlen(simCmdList[cmdno].name);
+  simCmdList[cmdno].nlen = (int16_t)strlen(simCmdList[cmdno].name);
   simCmdList[cmdno].parseCmd = simParseSystemCmd;
   simCmdList[cmdno].executeCmd = simExecuteSystemCmd;
   simAddCmdIntoHash(&(simCmdList[cmdno]));
@@ -983,7 +983,7 @@ void simInitsimCmdList() {
   cmdno = SIM_CMD_SYSTEM_CONTENT;
   simCmdList[cmdno].cmdno = cmdno;
   strcpy(simCmdList[cmdno].name, "system_content");
-  simCmdList[cmdno].nlen = (int)strlen(simCmdList[cmdno].name);
+  simCmdList[cmdno].nlen = (int16_t)strlen(simCmdList[cmdno].name);
   simCmdList[cmdno].parseCmd = simParseSystemContentCmd;
   simCmdList[cmdno].executeCmd = simExecuteSystemContentCmd;
   simAddCmdIntoHash(&(simCmdList[cmdno]));
@@ -991,7 +991,7 @@ void simInitsimCmdList() {
   cmdno = SIM_CMD_PRINT;
   simCmdList[cmdno].cmdno = cmdno;
   strcpy(simCmdList[cmdno].name, "print");
-  simCmdList[cmdno].nlen = (int)strlen(simCmdList[cmdno].name);
+  simCmdList[cmdno].nlen = (int16_t)strlen(simCmdList[cmdno].name);
   simCmdList[cmdno].parseCmd = simParsePrintCmd;
   simCmdList[cmdno].executeCmd = simExecutePrintCmd;
   simAddCmdIntoHash(&(simCmdList[cmdno]));
@@ -999,7 +999,7 @@ void simInitsimCmdList() {
   cmdno = SIM_CMD_SQL;
   simCmdList[cmdno].cmdno = cmdno;
   strcpy(simCmdList[cmdno].name, "sql");
-  simCmdList[cmdno].nlen = (int)strlen(simCmdList[cmdno].name);
+  simCmdList[cmdno].nlen = (int16_t)strlen(simCmdList[cmdno].name);
   simCmdList[cmdno].parseCmd = simParseSqlCmd;
   simCmdList[cmdno].executeCmd = simExecuteSqlCmd;
   simAddCmdIntoHash(&(simCmdList[cmdno]));
@@ -1007,7 +1007,7 @@ void simInitsimCmdList() {
   cmdno = SIM_CMD_SQL_ERROR;
   simCmdList[cmdno].cmdno = cmdno;
   strcpy(simCmdList[cmdno].name, "sql_error");
-  simCmdList[cmdno].nlen = (int)strlen(simCmdList[cmdno].name);
+  simCmdList[cmdno].nlen = (int16_t)strlen(simCmdList[cmdno].name);
   simCmdList[cmdno].parseCmd = simParseSqlErrorCmd;
   simCmdList[cmdno].executeCmd = simExecuteSqlErrorCmd;
   simAddCmdIntoHash(&(simCmdList[cmdno]));
@@ -1015,7 +1015,7 @@ void simInitsimCmdList() {
   cmdno = SIM_CMD_SQL_SLOW;
   simCmdList[cmdno].cmdno = cmdno;
   strcpy(simCmdList[cmdno].name, "sql_slow");
-  simCmdList[cmdno].nlen = (int)strlen(simCmdList[cmdno].name);
+  simCmdList[cmdno].nlen = (int16_t)strlen(simCmdList[cmdno].name);
   simCmdList[cmdno].parseCmd = simParseSqlSlowCmd;
   simCmdList[cmdno].executeCmd = simExecuteSqlSlowCmd;
   simAddCmdIntoHash(&(simCmdList[cmdno]));
@@ -1024,14 +1024,14 @@ void simInitsimCmdList() {
   cmdno = SIM_CMD_TEST;
   simCmdList[cmdno].cmdno = cmdno;
   strcpy(simCmdList[cmdno].name, "test");
-  simCmdList[cmdno].nlen = (int)strlen(simCmdList[cmdno].name);
+  simCmdList[cmdno].nlen = (int16_t)strlen(simCmdList[cmdno].name);
   simCmdList[cmdno].parseCmd = NULL;
   simCmdList[cmdno].executeCmd = simExecuteTestCmd;
 
   cmdno = SIM_CMD_RETURN;
   simCmdList[cmdno].cmdno = cmdno;
   strcpy(simCmdList[cmdno].name, "return");
-  simCmdList[cmdno].nlen = (int)strlen(simCmdList[cmdno].name);
+  simCmdList[cmdno].nlen = (int16_t)strlen(simCmdList[cmdno].name);
   simCmdList[cmdno].parseCmd = simParseReturnCmd;
   simCmdList[cmdno].executeCmd = simExecuteReturnCmd;
   simAddCmdIntoHash(&(simCmdList[cmdno]));
